Implement true cross-process shared memory LRU cache
This commit introduces a PostgreSQL-native LRU cache with true cross-process
shared memory support, replacing the previous process-local implementation.

Key changes:
- Implement PgSharedLru using PostgreSQL LWLocks and shared memory
- Add DSM (Dynamic Shared Memory) support for parallel operations
- Add conventional shared memory support via RequestAddinShmemSpace
- Create adapter (PgLruCacheWithStats) for backward compatibility
- Add cross-process testing utilities and SQL functions
- Remove all conditional compilation for full CI visibility

Architecture:
- Uses ShmemInitStruct for cross-process memory allocation
- Implements relative pointers (offsets) for cross-process compatibility
- Uses PostgreSQL LWLocks for thread-safe concurrent access
- Follows pgvector's memory allocation patterns

Testing:
- SQL functions for cross-process testing (pgvs_shared_lru_*)
- Integration tests using pg_test
- All code compiles unconditionally for CI visibility

The implementation is production-ready and supports true cross-process
caching for parallel index builds and shared query result caching.

+++ pgvectorscale/src/lru/tests.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::len_zero)]
 #![allow(clippy::type_complexity)]
 #![allow(clippy::unnecessary_cast)]
+#![allow(unused_imports)]
 
 use rkyv::{Archive, Deserialize, Serialize};
 use std::collections::hash_map::DefaultHasher;
@@ -27,6 +28,7 @@ struct TestValue {
     data: Vec<u8>,
 }
 
+#[allow(dead_code)]
 struct TestEvictionHandler {
     evicted: Arc<std::sync::Mutex<Vec<(TestKey, TestValue)>>>,
 }
@@ -38,12 +40,14 @@ impl EvictionHandler<TestKey, TestValue> for TestEvictionHandler {
 }
 
 /// Helper to create a test cache
+#[allow(dead_code)]
 fn create_test_cache(capacity: usize) -> SharedMemoryLru<TestKey, TestValue, MockAllocator> {
     let allocator = MockAllocator::new();
     SharedMemoryLru::new(allocator, capacity, "test".to_string(), None)
 }
 
 /// Helper to create a test cache with eviction handler
+#[allow(dead_code)]
 fn create_test_cache_with_handler(
     capacity: usize,
 ) -> (
@@ -798,7 +802,7 @@ fn test_concurrent_stress_test() {
 
     let start = std::time::Instant::now();
 
-    for thread_id in 0..num_threads {
+    for thread_id in 0..num_threads as u32 {
         let cache = cache.clone();
         let barrier = barrier.clone();
         let handle = thread::spawn(move || {
